% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsi_colours.R
\docType{data}
\name{bsi_palettes}
\alias{bsi_palettes}
\title{Palettes derived from BSI brand colours}
\format{
A named list of character vectors (hex colours).
}
\usage{
bsi_palettes
}
\description{
A named list of palettes used by [get_pal()] and [print_pal()].
}
\keyword{datasets}
