% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsi_colours.R
\name{get_pal}
\alias{get_pal}
\title{Select a BSI palette}
\usage{
get_pal(name)
}
\arguments{
\item{name}{Palette name. One of `names(bsi_palettes)`.}
}
\value{
A named character vector of hex colours.
}
\description{
Select a BSI palette
}
\examples{
get_pal("secondary_dark")
get_pal("classic")
get_pal("secondary_light")
}
