% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsi_colours.R
\name{print_pal}
\alias{print_pal}
\title{Plot a colour palette}
\usage{
print_pal(x, labels = c("index", "none"), ...)
}
\arguments{
\item{x}{Either a palette name (character length 1) or a vector of colours.}

\item{labels}{What to label each swatch with:
- `"none"`: no labels
- `"index"`: 1, 2, 3, ...}

\item{...}{Reserved for future use.}
}
\value{
Invisibly returns the colour vector.
}
\description{
Plot a colour palette
}
\examples{
print_pal("classic")
print_pal(get_pal("secondary_dark", n = 5))
}
