% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsi_colours.R
\name{print_pal}
\alias{print_pal}
\title{Plot a colour palette}
\usage{
print_pal(x, ...)
}
\arguments{
\item{x}{A vector of colours (e.g., hex codes).}

\item{...}{Reserved for future use.}
}
\value{
Invisibly returns `x`.
}
\description{
Plot a colour palette
}
